Reorg location to common area
