+++ cal_ratio_trainer/reporting/evaluation_utils.py
@@ -1,38 +1,9 @@
-from dataclasses import dataclass
 from pathlib import Path
-from typing import List, Optional
 
-from keras.models import load_model
-from keras import Model
 import numpy as np
-
 import pandas as pd
 
-
-@dataclass
-class TrainedModelData:
-    "Information about an epoch along with test data we can use"
-
-    # The input data for the model
-    x: List[np.ndarray]
-
-    # The expected results (labels) for the test data
-    y: pd.Series
-
-    # The LLP information (mH, mS, etc.) for the training data
-    z: pd.DataFrame
-
-    # Weights associated with the x, y, and z.
-    weights: List[np.ndarray]
-
-
-@dataclass
-class TrainedModel:
-    # The loaded model and weights
-    model: Model
-
-    def predict(self, x: TrainedModelData) -> np.ndarray:
-        return self.model.predict(x.x, verbose="0")
+from cal_ratio_trainer.common.trained_model import TrainedModelData
 
 
 def load_test_data(run_dir: Path) -> TrainedModelData:
@@ -59,15 +30,3 @@ def load_test_data(run_dir: Path) -> TrainedModelData:
         z=z_data,
         weights=list(weights.values()),
     )
-
-
-def load_trained_model(run_dir: Path, epoch: int) -> TrainedModel:
-    # Make sure the epoch is there.
-    assert run_dir / "keras" / f"final_model_weights_{epoch}.h5"
-
-    # Load the model that was written out
-    model = load_model(run_dir / "keras" / "final_model.keras")  # type: Optional[Model]
-    assert model is not None, "Failed to load model"
-    model.load_weights(run_dir / "keras" / f"final_model_weights_{epoch}.keras")
-
-    return TrainedModel(model=model)
